refactor: HomePage 폴더 위치 정리
